# openapi.json
{
  "openapi":"3.1.0",
  "info":{
    "title":"Herb Classifier API",
    "version":"v1.3"
  },
  "paths":{"/api/auth/register":{"post":{"tags":["auth"],"summary":"Register User","description":"The endpoint for registering a new user\n\nArgs:\n    username (str): the username of the user\n    email (str): the email of the user\n    password (str): the password of the user\n    is_admin (bool): whether the user is an admin\n\nReturns:\n    (UserInDB) The user that was registered\n\nRaises:\n    HTTPException: if the username already exists","operationId":"register_user_api_auth_register_post","parameters":[{"name":"is_admin","in":"query","required":false,"schema":{"default":false,"title":"Is Admin"}}],"requestBody":{"required":true,"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_register_user_api_auth_register_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"The requested page was not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["auth"],"summary":"Login For Access Token","description":"The endpoint for logging in a user\n\nArgs:\n    form_data (OAuth2PasswordRequestForm): the form data for the user\n\nReturns:\n    (dict) The access token for the user\n\nRaises:\n    HTTPException: if the username or password is incorrect","operationId":"login_for_access_token_api_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_api_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"The requested page was not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/logout":{"post":{"tags":["auth"],"summary":"Logout","description":"The endpoint for logging out a user\n\nArgs:\n    token (oauth2 bearer token): the token for the user\n\nReturns:\n    (dict) The message for logging out","operationId":"logout_api_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"The requested page was not found"}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/herb/search-herbs":{"post":{"tags":["herb"],"summary":"Search Herb By Disease","description":"The endpoint for searching a herb by disease\n\nArgs:\n    disease (str): the disease to search for\n    token (oauth2 bearer token): the token for the user\n\nReturns:\n    (str) The herb for the disease","operationId":"search_herb_by_disease_api_herb_search_herbs_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_search_herb_by_disease_api_herb_search_herbs_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"The requested page was not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/herb/search-diseases":{"post":{"tags":["herb"],"summary":"Search Herb By Disease","description":"The endpoint for searching a disease by herb\n\nArgs:\n    herb (str): the herb to search for\n    token (oauth2 bearer token): the token for the user\n\nReturns:\n    (str) The disease for the herb","operationId":"search_herb_by_disease_api_herb_search_diseases_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_search_herb_by_disease_api_herb_search_diseases_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"The requested page was not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/herb/evaluate":{"post":{"tags":["herb"],"summary":"Evaluate Image","operationId":"evaluate_image_api_herb_evaluate_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_evaluate_image_api_herb_evaluate_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"The requested page was not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/disease/insert":{"post":{"tags":["disease"],"summary":"Insert Disease","description":"The endpoint for searching a disease\n\nArgs:\n   disease (str): the disease to search for\n   token (oauth2 bearer token): the token for the user\n   symptoms (str): disease symptoms\n   treatment (str): treatment for the disease\n\nReturns:\n   (str) Successfully added new disease details record\n\nRaises:\n   HTTPException: if the user is not logged in","operationId":"insert_disease_api_disease_insert_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_insert_disease_api_disease_insert_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"The requested page was not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/disease/search":{"post":{"tags":["disease"],"summary":"Search Disease","description":"The endpoint for searching a disease\n\nArgs:\n    disease (str): the disease to search for\n    token (oauth2 bearer token): the token for the user\n\nReturns:\n    (str) The disease details\n\nRaises:\n    HTTPException: if the user is not logged in","operationId":"search_disease_api_disease_search_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_search_disease_api_disease_search_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"The requested page was not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/location/get-location":{"post":{"tags":["location"],"summary":"Get Location For Herb","description":"The endpoint for getting location for a herb\nArgs:\n    herb (str): the herb to search for\n    token (oauth2 bearer token): the token for the user\n\nReturns:\n    (str) The location for the herb\n\nRaises:\n    HTTPException: if the user is not logged in","operationId":"get_location_for_herb_api_location_get_location_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_get_location_for_herb_api_location_get_location_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"The requested page was not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/location/add-location":{"post":{"tags":["location"],"summary":"Add New Location","description":"The endpoint for adding a new location for a herb \nArgs:\n    lon (float): the longitude of the location\n    lat (float): the latitude of the location\n    herb (str): the herb to add the location for\n    token (oauth2 bearer token): the token for the user\n\nReturns:\n    (str) A message indicating the success of the operation\n\nRaises:\n    HTTPException: if the user is not logged in","operationId":"add_new_location_api_location_add_location_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_add_new_location_api_location_add_location_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"The requested page was not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_add_new_location_api_location_add_location_post":{"properties":{"lon":{"type":"number","title":"Lon"},"lat":{"type":"number","title":"Lat"},"herb":{"type":"string","title":"Herb"}},"type":"object","required":["lon","lat","herb"],"title":"Body_add_new_location_api_location_add_location_post"},"Body_evaluate_image_api_herb_evaluate_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"}},"type":"object","required":["image"],"title":"Body_evaluate_image_api_herb_evaluate_post"},"Body_get_location_for_herb_api_location_get_location_post":{"properties":{"herb":{"type":"string","title":"Herb"}},"type":"object","required":["herb"],"title":"Body_get_location_for_herb_api_location_get_location_post"},"Body_insert_disease_api_disease_insert_post":{"properties":{"disease":{"type":"string","title":"Disease"},"symptoms":{"type":"string","title":"Symptoms"},"treatment":{"type":"string","title":"Treatment"}},"type":"object","required":["disease","symptoms","treatment"],"title":"Body_insert_disease_api_disease_insert_post"},"Body_login_for_access_token_api_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_api_auth_login_post"},"Body_register_user_api_auth_register_post":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","email","password"],"title":"Body_register_user_api_auth_register_post"},"Body_search_disease_api_disease_search_post":{"properties":{"disease":{"type":"string","title":"Disease"}},"type":"object","required":["disease"],"title":"Body_search_disease_api_disease_search_post"},"Body_search_herb_by_disease_api_herb_search_diseases_post":{"properties":{"herb":{"type":"string","title":"Herb"}},"type":"object","required":["herb"],"title":"Body_search_herb_by_disease_api_herb_search_diseases_post"},"Body_search_herb_by_disease_api_herb_search_herbs_post":{"properties":{"disease":{"type":"string","title":"Disease"}},"type":"object","required":["disease"],"title":"Body_search_herb_by_disease_api_herb_search_herbs_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"http://127.0.0.1:8000/auth/login"}}}}}}